Debug

+++ RaspberryPi/CodeWithROS/communication/src/uartparser.cpp
@@ -24,7 +24,7 @@ UARTparser::~UARTparser() {
 
 void UARTparser::parseData(const communication::msgStruct msg) {	
 	// Search for sensor in IDmap that belongs to id of msg and parse data in specified function
-	std::map<uint32_t, Parser*>::iterator temp = IDmap.find(msg.id);
+	std::map<uint32_t, Parser*>::iterator temp = IDmap.find((uint32_t)msg.id);
 	
 	ROS_INFO("Parsing data, ID: 0x%X", msg.id);
